Added abstract seo configurator to hold common config tasks
