+++ Seo/Og/OgSeoConfigurator.php
@@ -2,32 +2,20 @@
 
 namespace Leogout\Bundle\SeoBundle\Seo\Og;
 
+use Leogout\Bundle\SeoBundle\Seo\AbstractSeoConfigurator;
+use Leogout\Bundle\SeoBundle\Seo\AbstractSeoGenerator;
+
 /**
  * Description of OgSeoConfigurator.
  *
  * @author: leogout
  */
-class OgSeoConfigurator
+class OgSeoConfigurator extends AbstractSeoConfigurator
 {
     /**
-     * @var array
+     * @param AbstractSeoGenerator $generator
      */
-    protected $config;
-
-    /**
-     * OgSeoConfigurator constructor.
-     *
-     * @param array $config
-     */
-    public function __construct(array $config)
-    {
-        $this->config = $config;
-    }
-
-    /**
-     * @param OgSeoGenerator $generator
-     */
-    public function configure(OgSeoGenerator $generator)
+    public function configure(AbstractSeoGenerator $generator)
     {
         if (null !== $title = $this->getConfig('title')) {
             $generator->setTitle($title);
@@ -44,19 +32,5 @@ class OgSeoConfigurator
         if (null !== $url = $this->getConfig('url')) {
             $generator->setImage($url);
         }
-    }
-
-    /**
-     * @param string $name
-     *
-     * @return string|null
-     */
-    private function getConfig($name)
-    {
-        if (!isset($this->config[$name])) {
-            return null;
-        }
-
-        return $this->config[$name];
     }
 }
